+++ src/modules/requests/requests.controller.ts
@@ -14,7 +14,7 @@ import { RequestsService } from './services/requests.service';
 import { JoiValidationPipe } from 'src/common/validator/joi-validation.pipe';
 import { RequestPayloadSchema } from './validator/request.schema-validator';
 import { QueryRequestDto } from './dtos/query-request.dto';
-import { AuthUser } from '../../common/interfaces/keycloak/keycloak-user.interface';
+import { AuthUser } from '../../common/interfaces/keycloak-user.interface';
 import { UserAccessService } from './../../common/providers/user-access.service';
 
 @Controller()
