+++ src/modules/requests/services/requests.service.ts
@@ -5,7 +5,7 @@ import { mapEntitytoInterface } from '../interfaces/response.interface';
 import { metaPagination, queryPagination } from 'src/common/helpers/pagination';
 import { ApiResponse } from '../../../common/interfaces/api-response.interface';
 import { QueryRequestDto } from '../dtos/query-request.dto';
-import { UserAccess } from '../../../common/interfaces/keycloak/keycloak-user.interface';
+import { UserAccess } from '../../../common/interfaces/keycloak-user.interface';
 
 @Injectable()
 export class RequestsService {
